Capture javac/jar output
Capture the output of compilation and packaging, and either return them as the
exception message on failure, or make them accessible to caller on the returned
object on success.  Successful messages include notes, warning, and the help
text (with --help/Options::setHelp).

Hides IOException from make(), but document ModuleCompilerException in
signature which is strictly unnecessary since it's a RuntimeException.

Improves test coverage.